chore(picker): Allow to set up any picker. Default vim.ui.select
